Tried Adding Upper Grounding Strip to Mirror
The upper grounding strip has been successfully added to the mirror
file, but for some reason there seems to be a material/physics domain
issue when opened in comsol. It appears the domains in the code are
correct since when manually entered in comsol everything computes. This
has to be looked at.

Also, work was started on creating the inner strips for the single strip
file. The geometry should be created but has not been checked in comsol,
and selections and physics still need to be done for this file.